chore: props provider instead of window
