+++ packages/f36-docs-utils/src/index.ts
@@ -1,1 +1,2 @@
 export { Props } from './Props/Props';
+export { PropsProvider } from './Props/PropsProvider';
