+++ packages/forma-36-website/gatsby-node.js
@@ -11,6 +11,7 @@ function getTypescriptMetaInformation(sourcePath) {
         savePropValueAsString: true,
         shouldExtractLiteralValuesFromEnum: true,
         shouldExtractValuesFromUnion: true,
+        skipChildrenPropWithoutDoc: false,
       },
     );
